pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "QSBets"
version = "0.10.0"
requires-python = ">=3.12, <3.14"
description = "Event-driven stock analysis and investment recommendation system"
dependencies = [
    "langchain-ollama",
    "langchain-azure-ai",
    # "langchain-chroma",
    "langchain_community",
    "python-dotenv",
    "numpy",
    "pydantic",
    "pandas",
    "selenium",
    "trafilatura",
    "azure-ai-inference",
    "edgartools",
    "ollama",
    "diskcache",
    "chromadb",
    "rich",
    "praw", # for redditor api but currently unused
    "ta-lib",
    "fredapi",
    "mlx-lm",
    "pyyaml",
    "streamlit",
    "plotly"
]

[project.scripts]
qsbets = "main:main"

[tool.hatch.build.targets.wheel]
packages = ["QSBets"]

[tool.poe.tasks]
test.sequence = [
    "pytest -n auto",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"